requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow
psutil
py-tgcalls==1.2.9
ntgcalls==1.1.2
Flask==2.0.3
gunicorn==20.1.0
pykeyboard
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
uvloop
yt-dlp
youtube-search
youtube-search-python
Flask
Flask-RESTful
pymongo
httpx==0.27.2
pyrogram
